ENT-4911: Split network verifier's sealed CorDapps into separate packages.
